vps/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
duckdb>=1.2.0
yfinance>=0.2.50
python-dotenv>=1.0.0
